% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_data_preparation.R
\name{get_data_butik_reference_unit}
\alias{get_data_butik_reference_unit}
\title{Generate data sets per butik/shop, week, and reference unit.}
\usage{
get_data_butik_reference_unit(data_all, data_meta, time_frq, ref_unit)
}
\arguments{
\item{data_all}{the overall data set (containing all possible refernce units)}

\item{data_meta}{the meta data as generated via \code{\link[=get_meta_data]{get_meta_data()}}}

\item{time_frq}{a character specifying the time frequency (either 'daily' or
'weekly')}

\item{ref_unit}{a character specifying the reference unit of either
\itemize{
\item "zoner"
\item "kampanj"
\item "skylttyp"
}}
}
\value{
a list of length number of weeks, with elements being lists of length
number of unique butiks which gives the overall data set as parts/slices
per week and butik
}
\description{
Applies filtering commands to daily or weekly data sets according to uniquely
available butiks in the given data, the reference unit (zoner, kampanj,
skylttyp) and subject to the time frequency.
}
