% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_data_preparation.R
\name{get_list_data_sales_prod_adv}
\alias{get_list_data_sales_prod_adv}
\title{Generates a list of data sets with advertised (target) products.}
\usage{
get_list_data_sales_prod_adv(vecka_seq, pth_data_prod_adv)
}
\arguments{
\item{vecka_seq}{integer vector giving the week numbers to use}

\item{pth_data_prod_adv}{a character vector of paths to the data sets (per
week) that have all IDs}
}
\value{
a list of data sets witch each element representing data (for a week)
that is cleaned for relevant product IDs
}
\description{
Output List elements are taken per week (vecka)
}
