% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hypothesis_tests.R
\name{get_matrix_fisher}
\alias{get_matrix_fisher}
\title{Generates numeric matrix from data subset suitable for \code{\link[=fisher.test]{fisher.test()}}.}
\usage{
get_matrix_fisher(data_sub_set, var_to_test, cntr_period, test_period)
}
\arguments{
\item{data_sub_set}{a data set as passed via \code{data_set} from
\code{\link[=perform_tests]{perform_tests()}} but additionally filtered for testing and control period}

\item{cntr_period}{an integer sequence (from to e.g. \code{28:33}) specifying
the weeks of the control period}

\item{test_period}{an integer sequence (from to e.g. \code{34:35}) specifying
the weeks of the testing period}
}
\value{
a matrix suitable for \code{\link[=fisher.test]{fisher.test()}} with appropriate
\code{dimnames}
}
\description{
Generates numeric matrix from data subset suitable for \code{\link[=fisher.test]{fisher.test()}}.
}
