% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hypothesis_tests.R
\name{get_test_matrix}
\alias{get_test_matrix}
\title{Generates numeric matrix from data subset suitable for different tests}
\usage{
get_test_matrix(
  data_sub_set,
  var_to_test,
  type_group_var,
  cntrl_group,
  test_group
)
}
\arguments{
\item{data_sub_set}{a data set as passed via \code{data_set} from
\code{\link[=perform_tests]{perform_tests()}} but additionally filtered for testing and controlgroup}

\item{type_group_var}{type_group_var of discriminating variable to fix test vs.
control groups: either vecka (comparing groups) or butik (comparing stores)
against each other}

\item{cntrl_group}{an integer sequence (from to e.g. \code{28:33}) specifying
the weeks of the control group or a character sequence specifying the butik
names of the control group}

\item{test_group}{an integer sequence (from to e.g. \code{34:35}) specifying
the weeks of the testing group (if \code{type_group} == "vecka") or
(if \code{type_group} == "butik"), a character sequence specifying the
butik names of the testing group}
}
\value{
a matrix suitable as first argument to these tests (see Descripiton)
}
\description{
Supported tests are \code{\link[=prop.test]{prop.test()}}, \code{\link[=fisher.test]{fisher.test()}}, ...
}
