% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_data_preparation.R
\name{hs_get_data_to_weekly}
\alias{hs_get_data_to_weekly}
\title{Read sales data on advertised products and split into test and control weeks}
\usage{
hs_get_data_to_weekly(pth_to_dataset, vecka_seq, cntrl_min, cntrl_max)
}
\arguments{
\item{pth_to_dataset}{pth to the overall data set}

\item{vecka_seq}{a vector of double digit integers specifying the test weeks}

\item{cntrl_min}{6 digit integer indicating the first control week (e.g.
202201)}

\item{cntrl_max}{6 digit integer indicating the last control week (e.g.
202233)}
}
\value{
side effect function writing the data sets to the "data/..." location
}
\description{
The data set need to be split because this is required input for all
subsequent operations which are based on data (for all butiks) in the test
week under consideration and control weeks 2022-01 to 2022-33, which are
passed via corresponding arguments. Also the 'HEMKÖP' part in the butik name
is removed.
}
