% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_data_preparation.R
\name{hs_merge_advertised_vecka}
\alias{hs_merge_advertised_vecka}
\title{Merge data sets of advertised products}
\usage{
hs_merge_advertised_vecka(pth_to_data, vecka_seq)
}
\arguments{
\item{pth_to_data}{path to data directory}

\item{vecka_seq}{integer sequence of advertised productID's data sets, where
each such data set contains control weeks and current 'vecka' week.}
}
\value{
a merged data set containing all productsID's from all kampanj
}
\description{
Advertised products are those under digital signage. Merging occurs for
different productID's. When duplicated productID's are identical, they are
merged.
}
