% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hypothesis_tests.R
\name{hs_perform_test}
\alias{hs_perform_test}
\title{Performs the exact Fisher test for the specific subset of data.}
\usage{
hs_perform_test(
  data_sub_set,
  var_to_test,
  type_test,
  type_group_var,
  cntrl_group,
  test_group,
  options = list(conf_level = 0.95, alternative = "two.sided")
)
}
\arguments{
\item{data_sub_set}{a data set as passed via \code{data_set} from
\code{\link[=hs_perform_tests]{hs_perform_tests()}} but additionally filtered for testing and controlgroup}

\item{type_test}{character value identifying the type of test to perform}

\item{type_group_var}{type_group_var of discriminating variable to fix test vs.
control groups: either vecka (comparing groups) or butik (comparing stores)
against each other}

\item{cntrl_group}{an integer sequence (from to e.g. \code{28:33}) specifying
the weeks of the control group or a character sequence specifying the butik
names of the control group}

\item{test_group}{an integer sequence (from to e.g. \code{34:35}) specifying
the weeks of the testing group (if \code{type_group} == "vecka") or
(if \code{type_group} == "butik"), a character sequence specifying the
butik names of the testing group}

\item{options}{options passed to test function as a list of two elements:
\itemize{
\item \code{conf_level}: defaults to 0.95 and within range (0,1)
\item \code{alternative}: defaults to "two.sided" with other options
being "greater" or "less"
}}
}
\value{
data matrix for, and output matrix from different tests e.g.
\code{\link[=fisher.test]{fisher.test()}} for a single specific data (subset)
}
\description{
Performs the exact Fisher test for the specific subset of data.
}
