% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hypothesis_tests.R
\name{hs_perform_tests}
\alias{hs_perform_tests}
\title{Performs various hypothesis tests on data.}
\usage{
hs_perform_tests(
  data_set,
  type_test,
  type_group_var = NULL,
  cntrl_group,
  test_group,
  options
)
}
\arguments{
\item{data_set}{the data set for a specific butik/shop}

\item{type_test}{character value identifying the type of test to perform}

\item{type_group_var}{type_group_var of discriminating variable to fix test vs.
control groups: either vecka (comparing groups) or butik (comparing stores)
against each other}

\item{cntrl_group}{an integer sequence (from to e.g. \code{28:33}) specifying
the weeks of the control group or a character sequence specifying the butik
names of the control group}

\item{test_group}{an integer sequence (from to e.g. \code{34:35}) specifying
the weeks of the testing group (if \code{type_group} == "vecka") or
(if \code{type_group} == "butik"), a character sequence specifying the
butik names of the testing group}

\item{options}{options passed to test function as a list of two elements:
\itemize{
\item \code{conf_level}: defaults to 0.95 and within range (0,1)
\item \code{alternative}: defaults to "two.sided" with other options
being "greater" or "less"
}}
}
\value{
a list of hypothesis-test outputs with each element being a test for
a sub-sample of the data (see internals for details) that contains the
test-output (which varies depending on \code{type_test})
}
\description{
Hypothesis tests performed consider the different proportions of sales
of advertised products between specified control group (\code{cntrl_group})
and testing group (\code{test_group}), the latter being characterized by
the placement of screens to enhance consumer demand and increase sales.
}
