% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_hypothesis_tests.R
\name{perform_tests}
\alias{perform_tests}
\title{Performs various hypothesis tests on data.}
\usage{
perform_tests(data_set, type = NULL, cntrl, test)
}
\arguments{
\item{data_set}{the data set for a specific butik/shop}

\item{type}{type of discriminating variable to fix test vs. control groups:
either vecka (comparing groups) or butik (comparing stores) against each
other}

\item{cntrl}{an integer sequence (from to e.g. \code{28:33}) specifying
the weeks of the control group or a character sequence specifying the butik
names of the control group}

\item{test}{an integer sequence (from to e.g. \code{34:35}) specifying
the weeks of the testing group or a character sequence specifying the butik
names of the testing group}
}
\value{
a list of hypothesis-test outputs of dimension
\code{length(ref_unit) + 1} with each element containing the corresponding
tests performed (Fisher, ..)
}
\description{
Hypothesis tests performed consider the different proportions of sales
of advertised products between specified control group (\code{cntrl})
and testing group (\code{test}), the latter being characterized by
the placement of screens to enhance consumer demand and increase sales.
}
