% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZetaAnalyticsTB.R
\docType{package}
\name{ZetaAnalyticsTB}
\alias{ZetaAnalyticsTB}
\title{ZetaAnalyticsTB:}
\description{
A package that comprises analytic tools developed in freelance projects for
Zeta Display.
}
\section{Introduction}{

If ever to be written, this sections outlines the package structure and gives
an overview of the facilities.
}

