% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_helper_data_preparation.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Generates list of zoner, kampanj and skylttyp for a all butiks and weeks.}
\usage{
get_meta_data(pth, data_validate)
}
\arguments{
\item{pth}{path to all data sets}
}
\value{
a list of lists, with element of the latter being output from
\code{\link[=get_list_test_week]{get_list_test_week()}}.
}
\description{
The butiks can probably be treated equally but the weeks will have different
products for zoner, kampanj or skylttyp so the product sets referring to
those need adjustments. The list summarizes the output for each week as
returned from \code{\link[=get_list_test_week]{get_list_test_week()}}.
}
